feat: New condition and extra field
Added a `variables` field for enchantments. Used to
store boilerplate strings used in arguments of actions
and conditions. Values from the `variables`
field will replace the variables inside the
arguments during the process of deserialization.

Added a `has-enchantment` condition, which expects
a seeker name and an enchantment key string.

This commit is not tested.